moved filter in modulator sub-package
