+++ core/modulator/simplefilter.go
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package core
+package modulator
 
 import (
 	"github.com/trivago/tgo/tlog"
+	"github.com/trivago/gollum/core"
 )
 
 // SimpleFilter plugin base type
@@ -32,7 +33,7 @@ import (
 // You can disable this behavior by setting "". Set to "" by default.
 type SimpleFilter struct {
 	Log          tlog.LogScope
-	dropStreamID MessageStreamID
+	dropStreamID core.MessageStreamID
 }
 
 // SetLogScope sets the log scope to be used for this filter
@@ -41,18 +42,18 @@ func (filter *SimpleFilter) SetLogScope(log tlog.LogScope) {
 }
 
 // Configure sets up all values requred by SimpleFormatter.
-func (filter *SimpleFilter) Configure(conf PluginConfigReader) error {
+func (filter *SimpleFilter) Configure(conf core.PluginConfigReader) error {
 	filter.Log = conf.GetSubLogScope("Filter")
 
-	filter.dropStreamID = GetStreamID(conf.GetString("DropToStream", InvalidStream))
+	filter.dropStreamID = core.GetStreamID(conf.GetString("DropToStream", core.InvalidStream))
 	return nil
 }
 
 // Drop sends the given message to the stream configured with this filter.
-func (filter *SimpleFilter) Drop(msg *Message) ModulateResult {
-	if filter.dropStreamID != InvalidStreamID {
+func (filter *SimpleFilter) Drop(msg *core.Message) core.ModulateResult {
+	if filter.dropStreamID != core.InvalidStreamID {
 		msg.SetStreamID(filter.dropStreamID)
-		return ModulateResultDrop
+		return core.ModulateResultDrop
 	}
-	return ModulateResultDiscard
+	return core.ModulateResultDiscard
 }
