+++ filter/sample.go
@@ -18,6 +18,7 @@ import (
 	"sync/atomic"
 
 	"github.com/trivago/gollum/core"
+	"github.com/trivago/gollum/core/modulator"
 )
 
 // Sample filter plugin
@@ -46,7 +47,7 @@ import (
 // sampling. This is useful for e.g. producers listeing to "*".
 // By default this list is empty.
 type Sample struct {
-	core.SimpleFilter
+	modulator.SimpleFilter
 	rate         int64
 	group        int64
 	count        *int64
